fix rating out of 10

+++ bookclub/models.py
@@ -221,7 +221,7 @@ class Rating(models.Model):
         blank = True
     )
 
-    rating = models.PositiveIntegerField(
+    rating = models.FloatField(
         blank=True,
         validators=[
             MaxValueValidator(10),
